Move main into RaidFinderGraph, add defaults
